% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_field.R
\name{clean_field}
\alias{clean_field}
\title{Clean a single metadata field}
\usage{
clean_field(x)
}
\arguments{
\item{x}{A vector (usually character/factor).}
}
\value{
A character vector.
}
\description{
Removes bracket wrappers like \code{[}xxx\verb{]} or \code{[}xxx\verb{]}, trims whitespace, and converts
common missing tokens to NA.
}
