% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_standardize.R
\name{country_unmatched_values}
\alias{country_unmatched_values}
\title{List unique unmatched country values}
\usage{
country_unmatched_values(x, ...)
}
\arguments{
\item{x}{A character vector (raw country/location strings).}

\item{...}{Passed to \code{standardize_country()}.}
}
\value{
A character vector of unique cleaned countries that failed to map.
}
\description{
List unique unmatched country values
}
