% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree_safe.R
\name{make_tree_safe}
\alias{make_tree_safe}
\title{Make strings safe for tree labels / filenames}
\usage{
make_tree_safe(
  x,
  na_token = "None",
  ascii = TRUE,
  tolower = FALSE,
  replacement = "_",
  allowed_chars_regex = "A-Za-z0-9._\\\\|-",
  collapse_replacement = TRUE,
  trim_replacement = TRUE
)
}
\arguments{
\item{x}{Character vector.}

\item{na_token}{Replacement for NA/empty.}

\item{ascii}{Logical; if TRUE, transliterate to ASCII (remove accents).}

\item{tolower}{Logical; if TRUE, convert to lower case.}

\item{replacement}{String used to replace invalid characters (default "_").}

\item{allowed_chars_regex}{Regex character class for allowed characters
(default keeps A-Z a-z 0-9 plus "._|-").}

\item{collapse_replacement}{Logical; collapse multiple replacements into one.}

\item{trim_replacement}{Logical; trim replacement at start/end.}
}
\value{
Character vector.
}
\description{
Make strings safe for tree labels / filenames
}
