% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbimeta-package.R
\docType{package}
\name{ncbimeta-package}
\alias{ncbimeta}
\alias{ncbimeta-package}
\title{ncbimeta: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
